test: satisfy reconciliation quality gates
